Adjusted security config structure, adjusted memory
For security config changed it to be a file instead of a folder. Changed memory to retrieve an ID of type Integer instead of Long

+++ src/main/java/com/timewave/timewave/repository/MemoryRepository.java
@@ -6,5 +6,5 @@ import java.util.List;
 
 public interface MemoryRepository extends JpaRepository<Memory, Long> {
     // Method to find memories by user
-    List<Memory> findByUserId(Long userId);
+    List<Memory> findByUserId(Integer userId);
 }
